refactor: move TimePath and TimePaths to domain package
