+++ src/main/java/nextstep/subway/maps/map/domain/path/TimePaths.java
@@ -1,4 +1,4 @@
-package nextstep.subway.maps.map.application.path;
+package nextstep.subway.maps.map.domain.path;
 
 import java.time.LocalDateTime;
 import java.util.Comparator;
