+++ src/test/java/nextstep/subway/maps/map/domain/path/TimePathsTest.java
@@ -1,4 +1,4 @@
-package nextstep.subway.maps.map.application.path;
+package nextstep.subway.maps.map.domain.path;
 
 import static org.assertj.core.api.Assertions.*;
