+++ src/test/java/nextstep/subway/utils/TestObjectUtils.java
@@ -1,10 +1,11 @@
 package nextstep.subway.utils;
 
-import nextstep.subway.maps.line.domain.Line;
-import nextstep.subway.maps.station.domain.Station;
+import java.time.LocalTime;
+
 import org.springframework.test.util.ReflectionTestUtils;
 
-import java.time.LocalTime;
+import nextstep.subway.maps.line.domain.Line;
+import nextstep.subway.maps.station.domain.Station;
 
 public class TestObjectUtils {
     public static Station createStation(Long id, String name) {
@@ -13,8 +14,10 @@ public class TestObjectUtils {
         return station;
     }
 
-    public static Line createLine(Long id, String name, String color, int overFare) {
-        Line line1 = new Line(name, color, LocalTime.of(5, 30), LocalTime.of(23, 30), 10, overFare);
+    public static Line createLine(Long id, String name, String color, int overFare, int intervalTime) {
+        Line line1 = new Line(
+            name, color, LocalTime.of(5, 30), LocalTime.of(23, 30), intervalTime, overFare
+        );
         ReflectionTestUtils.setField(line1, "id", id);
         return line1;
     }
